Add project files

+++ src/user/services/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { UserRepository } from '../repositories/user.repository';
 import { User } from 'models/user.model';
 import { UpdateUserType } from '../utils/types';
@@ -16,16 +16,30 @@ export class UserService {
     async getById(id: string): Promise<User> {
         const user = await this.userRepository.getUserById(id);
 
+        if (!user) {
+            throw new NotFoundException('User not found');
+        };
+
         return user;
     };
 
     async update(id: string, updateUserType: UpdateUserType): Promise<User> {
         const user = await this.userRepository.updateUser(id, updateUserType);
 
+        if (!user) {
+            throw new NotFoundException('User not found');
+        };
+
         return user;
     };
 
     async delete(id: string): Promise<boolean> {
+        const user = await this.userRepository.getUserById(id);
+
+        if (!user) {
+            throw new NotFoundException('User not found');
+        };
+        
         const result = await this.userRepository.deleteUser(id);
 
         return result;
